requirements.txt
Flask==3.1.2
flask-cors==6.0.1
supabase==2.24.0
python-dotenv==1.2.1
PyJWT==2.10.1
bcrypt==5.0.0
requests==2.32.5
httpx==0.28.1
websockets>=13.0,<16.0

